Refactor: move aligning methods to ContentManager
